DiscoRape/requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorthief==0.2.1
decorator==4.4.2
discord.py==1.5.0
emoji==0.6.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
mtranslate==1.6
multidict==4.7.6
mypy-extensions==0.4.3
nekos.py==1.0.3
numpy==1.19.2
pathspec==0.8.0
Pillow==8.0.0
proglog==0.1.9
psutil==5.7.2
pyfiglet==0.8.post1
regex==2020.10.15
requests==2.24.0
soupsieve==2.0.1
toml==0.10.1
tqdm==4.50.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
yarl==1.5.1